Use timezones of commit times and `since` strings
The commit dates of tags might not be in UTC, but other parts of the
code expect them to be.

Instead of touching multiple methods in `lib.py` to correctly handle non
UTC `datetime` objects it is simpler to always convert them to UTC in
`get_datetime_and_type`.

When a user manually enters a `since` datetime, this also needs to be
normalised to UTC.  If the string doesn't contain a UTC offset, then
local time is assumed.

Fixes #20